# config/swaggerDocs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Medoc Backend API",
    "version": "1.0.0",
    "description": "API документация для Medoc Backend"
  },
  "servers": [
    {
      "url": "https://medoc.vastness.ru"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/documents/send": {
      "post": {
        "tags": [
          "Document"
        ],
        "summary": "Отправка документа на подписание",
        "description": "Отправляет документ на подписание определённому получателю.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "recipientName": {
                    "type": "string",
                    "example": "Петров Иван Александрович"
                  },
                  "recipientPhoneNumber": {
                    "type": "string",
                    "example": "71234567890"
                  },
                  "documentTitle": {
                    "type": "string",
                    "example": "Договор аренды"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Документ успешно отправлен",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "document": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "documentTitle": {
                          "type": "string"
                        },
                        "recipient": {
                          "type": "object",
                          "properties": {
                            "name": {
                              "type": "string"
                            },
                            "phoneNumber": {
                              "type": "string"
                            }
                          }
                        },
                        "sender": {
                          "type": "object",
                          "properties": {
                            "clinicName": {
                              "type": "string"
                            },
                            "name": {
                              "type": "string"
                            },
                            "phoneNumber": {
                              "type": "string"
                            }
                          }
                        },
                        "fileUrl": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Отсутствуют необходимые поля (recipientName, recipientPhoneNumber, documentTitle)"
          },
          "404": {
            "description": "Файл с указанным названием не найден"
          },
          "403": {
            "description": "Клиника не авторизована"
          },
          "500": {
            "description": "Ошибка сервера"
          }
        }
      }
    },
    "/api/documents/upload-and-send": {
      "post": {
        "tags": [
          "Document"
        ],
        "summary": "Загрузка файла и отправка на подписание",
        "description": "Загружает файл в хранилище и создаёт запись в базе данных для отправки на подписание.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Файл, который будет отправлен."
                  },
                  "documentTitle": {
                    "type": "string",
                    "example": "Договор аренды"
                  },
                  "recipientName": {
                    "type": "string",
                    "example": "Петров Иван Александрович"
                  },
                  "recipientPhoneNumber": {
                    "type": "string",
                    "example": "71234567890"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Файл загружен и документ создан для подписания"
          },
          "400": {
            "description": "Файл отсутствует в запросе"
          },
          "403": {
            "description": "Клиника не авторизована"
          },
          "500": {
            "description": "Ошибка при загрузке файла или создании документа"
          }
        }
      }
    },
    "/api/documents/delete/{documentId}": {
      "delete": {
        "tags": [
          "Document"
        ],
        "summary": "Удаление документа",
        "description": "Удаляет документ со статусами 'Отправлен' или 'Отклонён'.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "documentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Документ успешно удалён",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "documentTitle": {
                      "type": "string"
                    },
                    "dateSigned": {
                      "type": "string",
                      "format": "date-time",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Документ нельзя удалить (он уже подписан)"
          },
          "404": {
            "description": "Документ не найден"
          },
          "403": {
            "description": "У вас нет прав на удаление этого документа"
          },
          "500": {
            "description": "Ошибка сервера"
          }
        }
      }
    },
    "/api/documents/sent-documents": {
      "get": {
        "tags": [
          "Document"
        ],
        "summary": "Получение отправленных документов",
        "description": "Возвращает список отправленных документов с фильтрацией и пагинацией.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "2025-01-01"
            },
            "description": "Фильтр по дате создания (начиная с). Формат YYYY-MM-DD"
          },
          {
            "name": "endDate",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "2025-12-31"
            },
            "description": "Фильтр по дате создания (до). Формат YYYY-MM-DD"
          },
          {
            "name": "status",
            "in": "query",
            "explode": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "docSent",
                  "docRejected",
                  "docSigned"
                ]
              }
            },
            "description": "Фильтр по статусам: docSent = Отправлен, docRejected = Отклонён, docSigned = Подписан."
          },
          {
            "name": "recipientName",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Поиск по ФИО получателя (регистронезависимый)"
          },
          {
            "name": "recipientPhoneNumber",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Поиск по телефону получателя (регистронезависимый)"
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Номер страницы"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Количество документов на страницу"
          }
        ],
        "responses": {
          "200": {
            "description": "Список документов с пагинацией",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "documents": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "_id": {
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          },
                          "recipient": {
                            "type": "object"
                          },
                          "sender": {
                            "type": "object"
                          },
                          "fileUrl": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "dateSigned": {
                            "type": "string",
                            "format": "date-time",
                            "nullable": true
                          }
                        }
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "number"
                        },
                        "page": {
                          "type": "number"
                        },
                        "limit": {
                          "type": "number"
                        },
                        "totalPages": {
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Ошибка сервера"
          }
        }
      }
    },
    "/api/documents/for-patient": {
      "get": {
        "tags": [
          "Document"
        ],
        "summary": "Получение документов для пациента",
        "description": "Возвращает документы для пациента, сгруппированные по клиникам.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "clinicName",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Фильтр по названию клиники"
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список документов, сгруппированных по клиникам",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "groupedDocuments": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "_id": {
                              "type": "string"
                            },
                            "title": {
                              "type": "string"
                            },
                            "sender": {
                              "type": "object"
                            },
                            "fileUrl": {
                              "type": "string"
                            },
                            "status": {
                              "type": "string"
                            },
                            "createdAt": {
                              "type": "string",
                              "format": "date-time"
                            }
                          }
                        }
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "number"
                        },
                        "page": {
                          "type": "number"
                        },
                        "limit": {
                          "type": "number"
                        },
                        "totalPages": {
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Ошибка сервера"
          }
        }
      }
    },
...
    "/api/contractors/{contractorId}": {
      "delete": {
        "tags": [
          "Contractor"
        ],
        "summary": "Удаление контрагента",
        "description": "Удаляет контрагента, если у него нет связанных документов (пустой documentIds).",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "contractorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Идентификатор контрагента (Contractor._id)"
          }
        ],
        "responses": {
          "200": {
            "description": "Контрагент успешно удалён"
          },
          "400": {
            "description": "Контрагента нельзя удалить, если у него есть документы"
          },
          "403": {
            "description": "Клиника не авторизована"
          },
          "404": {
            "description": "Контрагент не найден"
          },
          "500": {
            "description": "Ошибка сервера"
          }
        }
      }
    },
...
  }
}